Work around Pydantic issues #1268 and #2762
Pydantic's BaseModel.parse_obj currently accepts iterables of pairs and
converts them to dictionaries, but fails with a TypeError if any of the
resulting keys are not strings. Since I only expected ValidationError,
that exception propagated out.

A remote client can trigger this code path with an input like [[null,
null]], as Hypothesis discovered for me. Requiring the result of
json.loads to be a dictionary enforces that only string keys will be
passed to parse_obj, and also prevents some "creative" (undesirable)
encodings of valid input.

In addition, an input like {"__pydantic_self__": null} triggers a
similar TypeError. In case there are more TypeErrors I haven't found,
I'm not checking specifically for that case, but instead catching any
TypeError.